- refactor platforms tests - store platform name - add test for random platform selection
